Redone command names, help changes, get-saves-list command
Defauld command names are now verbose and hyphenated.
New aliases added for ease of use.
Help is in a transitional (broken) state in anticipation of an alias system refactor.
The get-saves-list command is seperated from load-memory and load-memory now loads the default save (_default.json) if no filename argument is passed.